+++ example/components/viewport.js
@@ -1,0 +1,35 @@
+import React, { Component } from 'react'
+import { Map, TileLayer } from '../../src'
+
+const DEFAULT_VIEWPORT = {
+  center: [51.505, -0.09],
+  zoom: 13,
+}
+
+export default class ViewportExample extends Component {
+  state = {
+    viewport: DEFAULT_VIEWPORT,
+  }
+
+  onClickReset = () => {
+    this.setState({ viewport: DEFAULT_VIEWPORT })
+  }
+
+  onViewportChanged = viewport => {
+    this.setState({ viewport })
+  }
+
+  render() {
+    return (
+      <Map
+        onClick={this.onClickReset}
+        onViewportChanged={this.onViewportChanged}
+        viewport={this.state.viewport}>
+        <TileLayer
+          attribution="&amp;copy <a href=&quot;http://osm.org/copyright&quot;>OpenStreetMap</a> contributors"
+          url="http://{s}.tile.osm.org/{z}/{x}/{y}.png"
+        />
+      </Map>
+    )
+  }
+}