+++ src/propTypes/viewport.js
@@ -1,0 +1,10 @@
+// @flow
+
+import PropTypes from 'prop-types'
+
+import latlng from './latlng'
+
+export default PropTypes.shape({
+  center: latlng,
+  zoom: PropTypes.number,
+})